Fix issue363:Unable to extract large 7z files. Properly changing a variable type up to uint64_t.

+++ libarchive/archive_read_support_format_7zip.c
@@ -1663,7 +1663,7 @@ parse_7zip_uint64(struct archive_read *a, uint64_t *val)
 			mask >>= 1;
 			continue;
 		}
-		*val += (avail & (mask -1)) << (8 * i);
+		*val += ((uint64_t)(avail & (mask -1))) << (8 * i);
 		break;
 	}
 	return (0);
